Move suite-specific blackbox logic into fuse suite
