+++ blackbox/harness/capabilities.py
@@ -4,10 +4,10 @@ import os
 import platform
 import shutil
 from pathlib import Path
-from typing import Any
+from typing import Any, Iterable
 
 
-def detect_capabilities() -> dict[str, Any]:
+def detect_capabilities(extra_tools: Iterable[str] = ()) -> dict[str, Any]:
     system = platform.system()
     caps: dict[str, Any] = {
         "os": system,
@@ -15,7 +15,6 @@ def detect_capabilities() -> dict[str, Any]:
         "python": platform.python_version(),
         "is_root": hasattr(os, "geteuid") and os.geteuid() == 0,
         "tools": {},
-        "fuse": {"ok": False, "detail": ""},
         "features": {
             "symlink": hasattr(os, "symlink"),
             "hardlink": hasattr(os, "link"),
@@ -26,48 +25,25 @@ def detect_capabilities() -> dict[str, Any]:
             "macos": system == "Darwin",
         },
     }
-    for tool in (
+    tools = {
         "bash",
         "docker",
         "podman",
-        "fio",
         "git",
-        "java",
         "make",
-        "mdtest",
-        "perl",
-        "prove",
         "rg",
-        "runltp",
-        "umount",
-        "vdbench",
-    ):
+        *extra_tools,
+    }
+    for tool in sorted(tools):
         caps["tools"][tool] = shutil.which(tool) or ""
-    caps["fuse"] = detect_fuse(system)
     caps["features"]["case_sensitive_host_tmp"] = detect_case_sensitive(Path(os.environ.get("TMPDIR", "/tmp")))
     return caps
-
-
-def detect_fuse(system: str) -> dict[str, str | bool]:
-    if system == "Linux":
-        if not Path("/dev/fuse").exists():
-            return {"ok": False, "detail": "/dev/fuse is missing"}
-        if not shutil.which("fusermount3") and not shutil.which("fusermount"):
-            return {"ok": False, "detail": "fusermount3/fusermount is missing"}
-        return {"ok": True, "detail": "Linux FUSE prerequisites are present"}
-    if system == "Darwin":
-        if shutil.which("mount_macfuse"):
-            return {"ok": True, "detail": "macFUSE mount helper is present"}
-        if shutil.which("mount_fusefs"):
-            return {"ok": True, "detail": "FUSE-T mount helper is present"}
-        return {"ok": False, "detail": "macFUSE/FUSE-T mount helper is missing"}
-    return {"ok": False, "detail": f"unsupported OS for FUSE suite: {system}"}
 
 
 def detect_case_sensitive(tmp_root: Path) -> bool | None:
     try:
         tmp_root.mkdir(parents=True, exist_ok=True)
-        probe = tmp_root / f"drive9-case-probe-{os.getpid()}"
+        probe = tmp_root / f"blackbox-case-probe-{os.getpid()}"
         probe.mkdir(exist_ok=True)
         upper = probe / "A"
         lower = probe / "a"
@@ -77,4 +53,4 @@ def detect_case_sensitive(tmp_root: Path) -> bool | None:
     except Exception:
         return None
     finally:
-        shutil.rmtree(tmp_root / f"drive9-case-probe-{os.getpid()}", ignore_errors=True)
+        shutil.rmtree(tmp_root / f"blackbox-case-probe-{os.getpid()}", ignore_errors=True)
