+++ blackbox/suites/fuse/modules/community_lock.py
@@ -16,7 +16,7 @@ class CommunityLock(BaseModule):
     def run(self, ctx: Context) -> dict[str, Any]:
         if not ctx.capabilities.get("features", {}).get("fcntl_lock"):
             raise ModuleSkip("fcntl locks are not supported on this platform", "platform skip")
-        remote = ctx.remote_root(self.id)
+        remote = ctx.target.remote_root(self.id)
         ctx.target.mkdir_remote(remote)
         handle = ctx.target.mount("community_lock", remote, durability="write-sync")
         try:
