+++ blackbox/suites/fuse/modules/community_mdtest.py
@@ -16,7 +16,7 @@ class CommunityMdtest(BaseModule):
 
     def run(self, ctx: Context) -> dict[str, Any]:
         mdtest = ctx.deps.ensure_mdtest()
-        remote = ctx.remote_root(self.id)
+        remote = ctx.target.remote_root(self.id)
         ctx.target.mkdir_remote(remote)
         handle = ctx.target.mount("community_mdtest", remote, durability="interactive")
         try:
