+++ blackbox/suites/fuse/modules/drive9_workflow_git_worktree.py
@@ -12,7 +12,7 @@ class Drive9GitWorktree(Drive9WorkflowBase):
 
     def run(self, ctx: Context) -> dict[str, Any]:
         bare = ctx.target.create_git_fixture("git-worktree-fixture")
-        remote = ctx.remote_root(self.id)
+        remote = ctx.target.remote_root(self.id)
         ctx.target.mkdir_remote(remote)
         handle = ctx.target.mount("drive9_git_worktree", remote, profile="coding-agent")
         try:
