+++ blackbox/suites/fuse/modules/ported_juicefs_random_rw.py
@@ -19,7 +19,7 @@ class PortedJuiceFSRandomRW(BaseModule):
         cfg = module_config(ctx, self.id)
         size = int(cfg.get("size_bytes", 4 * 1024 * 1024))
         ops = int(cfg.get("ops_by_preset", {}).get(ctx.selected_preset or "daily", 256))
-        remote = ctx.remote_root(self.id)
+        remote = ctx.target.remote_root(self.id)
         ctx.target.mkdir_remote(remote)
         handle = ctx.target.mount("ported_juicefs_random_rw", remote, durability="write-sync")
         try:
